+++ db/access.php
@@ -16,7 +16,7 @@
 
 
 /**
- * Defines the capabilities for the Metacourse Link block
+ * Defines the capabilities for the upload enrolment methods plugin
  *
  * @package    tool_uploadenrolmentmethods
  * @copyright   2018 Eoin Campbell
@@ -40,6 +40,14 @@ $capabilities = array(
 
         'captype' => 'write',
         'contextlevel' => CONTEXT_BLOCK,
+        'archetypes' => array(
+            'manager' => CAP_ALLOW
+        )
+    ),
+
+    'tool/uploadenrolmentmethods:use' => array(
+        'captype' => 'write',
+        'contextlevel' => CONTEXT_SYSTEM,
         'archetypes' => array(
             'manager' => CAP_ALLOW
         )